L10n scripts fixes.
